requirements.txt
annotated-types==0.7.0
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
cytoolz==1.0.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
filelock==3.17.0
hexbytes==1.3.0
hyperliquid-python-sdk==0.10.1
idna==3.10
msgpack==1.1.0
numpy==1.22.4
pandas==2.2.3
pandas-ta==0.3.14b0
parsimonious==0.10.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
